+++ src/ByteDev.FormUrlEncoded/PropertyInfoListExtensions.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+
+namespace ByteDev.FormUrlEncoded
+{
+    internal static class PropertyInfoListExtensions
+    {
+        public static PropertyInfo GetByAttributeName(this List<PropertyInfo> source, string name)
+        {
+            return source.SingleOrDefault(p => p.GetAttributeName() == name);
+        }
+    }
+}